Moved legacy channel mapping services and tools in their own plane
